+ Test 36, evaluating number of frames are read correctly + Moved some test functions from base.py -> verify.py
