+++ test/base.py
@@ -1,6 +1,7 @@
 import unittest
 import numpy as np
 import warnings
+import test.verify as verify
 from test.zipload import Zipload
 
 
@@ -39,86 +40,6 @@ class Base(unittest.TestCase):
                         value = p.string_value
                 print(fmt.format(g, p, value))
 
-
-    def compare_header(test_label, areader, breader, alabel, blabel, areal, breal):
-        ''' Ensure headers for data in reader a & b are equivalent.
-        '''
-
-        aheader = areader.header
-        bheader = breader.header
-        assert aheader.parameter_block == bheader.parameter_block, \
-            '{}, parameter_block: {} {}, {} {}'.format(test_label,
-            alabel, aheader.parameter_block, blabel, bheader.parameter_block)
-        assert aheader.parameter_block == bheader.parameter_block, \
-            '{}, data_block: {} {}, {} {}'.format(test_label,
-            alabel, aheader.data_block, blabel, bheader.data_block)
-        assert aheader.point_count == bheader.point_count, \
-            '{}, point_count: {} {}, {} {}'.format(test_label,
-            alabel, aheader.point_count, blabel, bheader.point_count)
-        assert aheader.analog_count == bheader.analog_count, \
-            '{}, analog_count: {} {}, {} {}'.format(test_label,
-            alabel, aheader.point_count, blabel, bheader.point_count)
-        assert aheader.first_frame == bheader.first_frame, \
-            '{}, point_count: {} {}, {} {}'.format(test_label,
-            alabel, aheader.first_frame, blabel, bheader.first_frame)
-        assert aheader.last_frame == bheader.last_frame, \
-            '{}, last_frame: {} {}, {} {}'.format(test_label,
-            alabel, aheader.last_frame, vbheader.last_frame)
-        assert aheader.analog_per_frame == bheader.analog_per_frame, \
-            '{}, analog_per_frame: {} {}, {} {}'.format(test_label,
-            alabel, aheader.analog_per_frame, blabel, bheader.analog_per_frame)
-        assert aheader.frame_rate == bheader.frame_rate, \
-            '{}, frame_rate: {} {}, {} {}'.format(test_label,
-            alabel, aheader.frame_rate, blabel, bheader.frame_rate)
-        assert aheader.max_gap == bheader.max_gap, \
-            '{}, max_gap: {} {}, {} {}'.format(test_label,
-            alabel, aheader.max_gap, blabel, bheader.max_gap)
-        assert aheader.max_gap == bheader.max_gap, \
-            '{}, max_gap: {} {}, {} {}'.format(test_label,
-            alabel, aheader.max_gap, blabel, bheader.max_gap)
-        assert aheader.long_event_labels == bheader.long_event_labels, \
-            '{}, long_event_labels: {} {}, {} {}'.format(test_label,
-            alabel, aheader.long_event_labels, blabel, bheader.long_event_labels)
-
-        event_mismatch = max(0, bheader.event_count - aheader.event_count)
-        for i in range(aheader.event_count):
-            event_label = aheader.event_labels[i]
-            if event_label in bheader.event_labels:
-                lindex = bheader.event_labels.index(event_label)
-                assert aheader.event_disp_flags[i] == bheader.event_disp_flags[lindex], \
-                    '{}, event_disp_flag {}: {} {}, {} {}'.format(test_label, event_label,
-                    alabel, aheader.event_disp_flags[i], blabel, bheader.event_disp_flags[lindex])
-                assert np.isclose(aheader.event_timings[i], bheader.event_timings[lindex], atol=1e-7), \
-                    '{}, event_timings {}: {} {}, {} {}'.format(test_label, event_label,
-                    alabel, aheader.event_timings[i], blabel, bheader.event_timings[lindex])
-            else:
-                event_mismatch += 1
-
-        # Validate event differences, up to a few labels can be mismatched
-        err_str = '''{}, label missmatch in event_block. File {} had {} and {} had {} with no match'''.format(
-            test_label, alabel, [l for l in aheader.event_labels if l not in bheader.event_labels],
-            blabel, [l for l in bheader.event_labels  if l not in aheader.event_labels])
-        if event_mismatch > 0:
-            if event_mismatch <= 2:
-                warnings.warn(err_str, ImportWarning)
-            else:
-                assert False, err_str
-
-
-        if areal:
-            assert aheader.scale_factor < 0.0, \
-                '{}, {} scale_factor: expected {} < 0'.format(test_label, alabel, aheader.scale_factor)
-        else:
-            assert aheader.scale_factor > 0.0, \
-                '{}, {} scale_factor: expected 0 < {}'.format(test_label, alabel, aheader.scale_factor)
-        if breal:
-            assert bheader.scale_factor < 0.0, \
-                '{}, {} scale_factor: expected {} < 0'.format(test_label, blabel, bheader.scale_factor)
-        else:
-            assert bheader.scale_factor > 0.0, \
-                '{}, {} scale_factor: expected 0 < {}'.format(test_label, blabel, bheader.scale_factor)
-
-
     def load_data(reader):
         ''' Fetch point and analog data arrays
         '''
@@ -155,7 +76,7 @@ class Base(unittest.TestCase):
         return point_frames[:, :, 4] >= 0
 
 
-    def compare_data(areader, breader, alabel, blabel):
+    def data_is_equal(areader, breader, alabel, blabel):
         ''' Ensure data in reader a & b are equivalent.
         '''
         # Number of scale factors data can deviate
@@ -218,45 +139,15 @@ class Base(unittest.TestCase):
             'Mismatched analog samples between {} and {}, number of sampled diff: {} of {}'.format(
                 alabel, blabel, analog_diff, nsampled_analog)
 
-    def array_has_values(array):
-        ''' Returns true if at least one dimension in the array is 0
-        '''
-        return np.prod(np.shape(array)) > 0
-
-    def verify_array_match_headers(point, analog, reader, label):
-        ''' Check point and analog arrays as fetched from load_data(reader), matches the headers in the reader.
-        '''
-        assert reader.point_used == np.shape(point)[1],\
-            'Mismatch in number of POINT samples for file {}, read {} expected {}'.format(
-             np.shape(point)[1], reader.point_used)
-        assert reader.analog_used == np.shape(analog)[1],\
-            'Mismatch in number of ANALOG samples for file {}, read {} expected {}'.format(
-             np.shape(analog)[1], reader.analog_used)
-
-    def count_nan(array):
-        return np.sum(np.isnan(array))
 
     def check_data_in_range(reader, label, min_range, max_range):
         point, analog = Base.load_data(reader)
         # Probably redundant but can be useful to verify
-        Base.verify_array_match_headers(point, analog, reader, label)
-
-        def values_in_range(array, min_range=min_range, max_range=max_range):
-            ''' Returns true if values in the array are within the (min, max) range.
-            '''
-            return np.all((min_range < array) & (array < max_range))
-
-        # Check POINT data
+        verify.array_match_headers(point, analog, reader, label)
+        # Check POINT data blocks
         if reader.point_used > 0:
-            # Mask in relation to camera bits
             point_masked = point[Base.create_camera_mask(point)]
-            assert Base.array_has_values(point_masked), 'No registered camera bits for file {}'.format(label)
-            for i in range(5):
-                assert values_in_range(point_masked[:, i]),\
-                        "POINT data for column {} was not in range ({}, {}) for file '{}'. Was in range ({}, {})".format(
-                        i, min_range, max_range, label, np.min(point_masked), np.max(point_masked))
-        # Check ANALOG data
+            verify.point_data_in_range(point_masked, label, min_range, max_range)
+        # Check ANALOG data blocks
         if reader.analog_used > 0:
-            assert values_in_range(analog),\
-                    "ANALOG data was not in range ({}, {}) for file '{}'. Was in range ({}, {})".format(
-                    min_range, max_range, label, np.min(analog), np.max(analog))
+            verify.analog_data_in_range(analog, label, min_range, max_range)
